Slightly reworked rewinding; added values rewinding.
Needs a solution to the problem of which values must be displayed
and/or interpolated. Currently, whenever the user rewinds, the
current error measure is displayed, regardless of what was being
displayed before. This will probably be trivial to solve once we
have a nice way of changing the current measure.

* Also changed all OpenMP-powered for loops to use signed integers,
  requirements of OMP2.x (which is what MSVC supports currently)
* The above change comes with a new header for utility functions